Adds conversion of models to ONNX.
Some tweaks have been made:
* "feature" from the CRAFT output is ignored.
* CRAFT output is reshaped back to BCHW.
* Input height to the recognition models have been made static and was
  set to the default config value (64).

Last one is important, as `AdaptiveAvgPool2d` is not converting properly
to ONNX. So the easiest way to fix that was to just assume the height
static and replace it with regular `AvgPool2d`.

Conversion of the "tamil_g1" still causes some errors, so it is skipped.